2022 WPI Update

+++ src/main/java/frc/robot/subsystems/TurretSubsystem.java
@@ -113,7 +113,7 @@ public class TurretSubsystem extends SubsystemBase {
     return isLinedUp;
   }
 
-  public int getEncoderPos() {
+  public double getEncoderPos() {
     return turretMotor.getSelectedSensorPosition();
   }
